requirements.txt
addict==2.4.0
antlr4-python3-runtime==4.9.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.0.1
contourpy==1.2.0
cuda-python==11.8.2
cupy-cuda11x==13.0.0
cxxfilt==0.3.0
cycler==0.12.1
Cython==3.0.9
dask==2024.1.1
dask-cuda==24.2.0
distributed==2024.1.1
faiss-gpu==1.7.2
fastrlock==0.8.2
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
fvcore==0.1.5.post20221221
h5py==3.11.0
idna==3.6
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
iopath==0.1.10
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.4
lit==18.1.2
llvmlite==0.42.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mmcls==0.25.0
mmcv-full==1.5.0
mmsegmentation==0.27.0
mpmath==1.3.0
msgpack==1.0.8
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.59.1
numpy==1.24.4
nvtx==0.2.10
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==24.0
pandas==1.5.3
partd==1.4.1
pillow==10.2.0
platformdirs==4.2.0
portalocker==2.8.2
prettytable==3.10.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==14.0.2
pycparser==2.22
Pygments==2.17.2
pynvml==11.4.1
pyparsing==3.1.2
pyre-extensions==0.0.23
python-dateutil==2.9.0.post0
python_papi==6.0.0.1
pytz==2024.1
PyYAML==6.0.1
rapids-dask-dependency==24.2.0
requests==2.31.0
rich==13.7.1
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.12.0
six==1.16.0
sortedcontainers==2.4.0
submitit==1.5.1
sympy==1.12
tabulate==0.9.0
tblib==3.0.0
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.4.18
tomli==2.0.1
toolz==0.12.1
torchmetrics==0.10.3
tornado==6.4
tqdm==4.66.2
treelite==4.0.0
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
xformers==0.0.18
yacs==0.1.8
yapf==0.40.2
zict==3.0.0
zipp==3.18.1